mcp_server/requirements.txt
fastapi
uvicorn
pydantic
google-genai
PyYAML
jsonschema
